refactor(ui): standardize component file structure (#218)
Align the src/components directory on the directory-per-component
pattern with index.ts barrel exports:

- Add AuthSetup/index.ts (only component missing a barrel export)
- Add atoms/index.ts barrel re-exporting all 9 atoms
- Rename 7 index.tsx barrels to index.ts (pure re-exports, no JSX):
  ActivityFeed, Issues, MyPRs, Notifications, ReviewQueue, Sidebar,
  Workspace
- Update AuthSetup imports in App.tsx and Settings/GitHubSettings.tsx
  to use the new barrel path

Closes #218
