+++ src/components/AuthSetup/index.ts
@@ -1,0 +1,1 @@
+export { AuthSetup } from "./AuthSetup";
