+++ src/components/Issues/index.ts
